(graphql-backed-by-snap-pre-1-code-coverage) Add tests for CompositeSolidDefinition.solids field
Summary:
Discovered that this was an untested codepath later in the diff stack. This is unused in
dagit right now because we end up fetching all solids at pipeline root level and shuffling it
all client-side. However this is a perfectly reasonable API to support long term and we should have the flexibility

Test Plan: BK

Reviewers: alangenfeld, themissinghlink, sashank

Differential Revision: https://dagster.phacility.com/D2271